fix: fixed issues from review
