+++ src/utils/tokenRegistry.ts
@@ -25,7 +25,7 @@ import {
 } from './chain/antelope';
 import { Asset, ChainType, IAccount, IChain, IPrivateKey, IToken } from './chain/types';
 import { appStorage, assetStorage, keyStorage } from './StorageManager/setup';
-import { EosioUtil, IUser } from '@tonomy/tonomy-id-sdk';
+import { EosioUtil, IUser, IUserBase } from '@tonomy/tonomy-id-sdk';
 import settings from '../settings';
 import Debug from 'debug';
 
@@ -145,7 +145,7 @@ export async function getKeyOrNullFromChain(chainEntry: TokenRegistryEntry): Pro
     }
 }
 
-export async function getAccountFromChain(chainEntry: TokenRegistryEntry, user: IUser): Promise<IAccount> {
+export async function getAccountFromChain(chainEntry: TokenRegistryEntry, user?: IUser): Promise<IAccount> {
     const { chain, token } = chainEntry;
     const asset = await assetStorage.findAssetByName(chainEntry.token);
     const key = await getKeyFromChain(chainEntry);
@@ -162,9 +162,11 @@ export async function getAccountFromChain(chainEntry: TokenRegistryEntry, user: 
             await assetStorage.createAsset(new Asset(token, BigInt(0)), account);
         }
     } else {
+        if (!user) throw new Error('User is required for Antelope chain to get account name');
         debug(
             `getAccountFromChain() fetching Antelope account for ${chain.getName()}: ${await user.getAccountName()}, ${await key.exportPrivateKey()}`
         );
+
         account = await AntelopeAccount.fromAccountAndPrivateKey(
             chain as AntelopeChain,
             await user.getAccountName(),
@@ -214,3 +216,17 @@ export const findEthereumTokenByChainId = (chainId: string) => {
 export const eip155StringToChainId = (eip155String: string) => {
     return eip155String.split(':')[1];
 };
+
+export async function addNativeTokenToAssetStorage(user: IUserBase) {
+    const accountName = await user.getAccountName();
+    const privateKey = await getKeyFromChain(pangeaTokenEntry);
+
+    const asset = new Asset(pangeaTokenEntry.token, BigInt(0));
+    const account = AntelopeAccount.fromAccountAndPrivateKey(
+        pangeaTokenEntry.chain,
+        accountName,
+        privateKey as AntelopePrivateKey
+    );
+
+    await assetStorage.createAsset(asset, account);
+}
